refactor: 예외 처리 메세지 추가
