+++ src/main/java/lotto/domain/LottoNumbers.java
@@ -1,10 +1,14 @@
-package lotto;
+package lotto.domain;
 
-import static lotto.LottoConstants.LOTTO_NUMBER_COUNT;
-import static lotto.LottoConstants.LOTTO_NUMBER_LOWER_BOUND;
-import static lotto.LottoConstants.LOTTO_NUMBER_UPPER_BOUND;
+import static lotto.domain.LottoConstants.LOTTO_NUMBER_COUNT;
+import static lotto.domain.LottoConstants.LOTTO_NUMBER_LOWER_BOUND;
+import static lotto.domain.LottoConstants.LOTTO_NUMBER_UPPER_BOUND;
+import static lotto.exception.ExceptionMessages.INVALID_COUNT_EXCEPTION_MESSAGE;
+import static lotto.exception.ExceptionMessages.NOT_UNIQUE_EXCEPTION_MESSAGE;
+import static lotto.exception.ExceptionMessages.OUT_OF_BOUNDS_EXCEPTION_MESSAGE;
 
 import java.util.List;
+import lotto.exception.ExceptionMessages;
 
 public class LottoNumbers {
 
@@ -23,13 +27,13 @@ public class LottoNumbers {
 
     private void validateUniqueness(List<Integer> lottoNumbers) {
         if (lottoNumbers.stream().distinct().count() != LOTTO_NUMBER_COUNT) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException(NOT_UNIQUE_EXCEPTION_MESSAGE);
         }
     }
 
     private void validateCount(List<Integer> lottoNumbers) {
         if (lottoNumbers.size() != LOTTO_NUMBER_COUNT) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException(INVALID_COUNT_EXCEPTION_MESSAGE);
         }
     }
 
@@ -39,7 +43,7 @@ public class LottoNumbers {
 
     private void validateSingleRange(int number) {
         if (number < LOTTO_NUMBER_LOWER_BOUND || number > LOTTO_NUMBER_UPPER_BOUND) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException(OUT_OF_BOUNDS_EXCEPTION_MESSAGE);
         }
     }
